refactor: FE 북마크 편집
